+++ packages/timeline/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @starbeam/timeline
 
+## 0.5.2
+
+### Patch Changes
+
+- 719a6fe: Updated package export maps
+- Updated dependencies [719a6fe]
+  - @starbeam/debug@0.5.2
+  - @starbeam/peer@1.0.2
+  - @starbeam/verify@0.5.2
+
 ## 0.5.1
 
 ### Patch Changes